pyproject.toml
[project]
name = "butler-cal"
version = "0.5.0"
description = "Scrape UT Calendar and update Google Calendar"
authors = [
  { name = "Adam Lewis" }
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  "requests",
  "beautifulsoup4",
  "pytz",
  "google-api-python-client",
  "google-auth",
  "google-auth-oauthlib",
  "google-auth-httplib2",
  "loguru",
]

# optional dev dependencies
[project.optional-dependencies]
dev = [
  "pytest",
  "mypy",
]

[project.scripts]
update-calendar = "butler_cal.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/butler_cal"]
